Add support for minimum time between GpsFixAndSend and ReasonForFix
This new functionality allows you to skip the GpsFixAndSend logic for a
certain amount of time.

This can be practical if you use both the Fix Interval and movement
interval, but don't want to get a GPS fix and Send Data all the times.
The GPS and send data probably takes the most battery life.

So, let's say you have your movement fix interval set at 15 minutes,
and also your GPS fix at 30 minutes.

Now, consider what happens if the device moves constantly.
That would mean you'd do 2 GPS fixes each hour, as well as 4 fixes based
on the movement.
That's 6 fixes, while you would have been perfectly happy with 4 fixes
as well - the movement fixes would be enough.

This new mfi= setting will allow you to set it to 15 minutes and
whenever the GpsFix needs to be performed, it will check if the previous
time was more than 15 minutes ago. If it was NOT, then it will not do a
GPS fix or send data.

The downside of this is that your maximum time between sends could
become double the amount of your normal GPS fix.
Let's say you have a movement fix, and then just before the MFI setting,
the normal GPS fix triggers. The GPS fix won't be send out, because the
movement was just done. But then the device does not move anymore. It
will take another normal GPS fix time before it tries again.

Because it's now less regular on when it sends out data, we have also
added a small flag (1 byte) that indicates the reason for the GPS fix
and send. This could help detemine if the delay is normal or if there's
something wrong with the device.